+++ R/manipulate_species.R
@@ -229,6 +229,7 @@ addSpecies <- function(params, species_params,
     p@maturity[old_sp, old_w] <- params@maturity
     p@sc[old_w] <- params@sc
     p@mu_b[old_sp, old_w] <- params@mu_b
+    p@ext_encounter[old_sp, old_w] <- params@ext_encounter
     p@intake_max[old_sp, old_w] <- params@intake_max
     p@search_vol[old_sp, old_w] <- params@search_vol
     p@metab[old_sp, old_w] <- params@metab
@@ -369,6 +370,7 @@ removeSpecies <- function(params, species) {
     }
     p@ft_mask <- params@ft_mask[keep, , drop = FALSE]
     p@mu_b <- params@mu_b[keep, , drop = FALSE]
+    p@ext_encounter <- params@ext_encounter[keep, , drop = FALSE]
     p@species_params <- p@species_params[keep, , drop = FALSE]
     p@interaction <- params@interaction[keep, keep, drop = FALSE]
     p@selectivity <- params@selectivity[, keep, , drop = FALSE]
@@ -460,6 +462,7 @@ renameSpecies <- function(params, replace) {
         dimnames(params@pred_kernel)$sp <- species
     }
     dimnames(params@mu_b)$sp <- species
+    dimnames(params@ext_encounter)$sp <- species
     dimnames(params@interaction)$predator <- species
     dimnames(params@interaction)$prey <- species
     dimnames(params@selectivity)$sp <- species